# Conflicts:
#	src/Entity/Repertoire.php

+++ src/Controller/RepertoireController.php
@@ -2,6 +2,7 @@
 
 namespace App\Controller;
 
+use App\Entity\Groupe;
 use App\Entity\Repertoire;
 use App\Repository\RepertoireRepository;
 use App\Service\RepertoireService;
@@ -71,16 +72,36 @@ final class RepertoireController extends AbstractController
 
     #[IsGranted('ROLE_USER')]
     #[Route('/repertoire/corbeille', name: 'repertoire_corbeille')]
-    public function corbeille(RepertoireRepository $repertoireRepository, YamlFileRepository $yamlFileRepository): Response
+    #[Route('/groupe/{id}/repertoire/corbeille', name: 'repertoire_corbeille_groupe')]
+    public function corbeille(
+        RepertoireRepository $repertoireRepository,
+        YamlFileRepository $yamlFileRepository,
+        ?Groupe $groupe = null
+    ): Response
     {
-        $user = $this->getUser();
-        $repertoires = $repertoireRepository->findDeletedByUser($user);
-        $yamlFiles   = $yamlFileRepository->findDeletedByUser($user);
+        if ($groupe) {
+            $this->denyAccessUnlessGranted('GROUPE_EDIT', $groupe);
 
-        return $this->render('yaml_file/corbeilleRepertoire.html.twig', [
-            'repertoires' => $repertoires,
-            'yamlFiles'   => $yamlFiles,
-        ]);
+            $repertoires = $repertoireRepository->findDeletedByGroupe($groupe);
+            $yamlFiles   = $yamlFileRepository->findDeletedByGroupe($groupe);
+
+            return $this->render('yaml_file/corbeilleRepertoireGroupe.html.twig', [
+                'repertoires' => $repertoires,
+                'yamlFiles'   => $yamlFiles,
+                'groupe'      => $groupe
+            ]);
+
+        } else {
+            $user = $this->getUser();
+
+            $repertoires = $repertoireRepository->findDeletedByUser($user);
+            $yamlFiles   = $yamlFileRepository->findDeletedByUser($user);
+
+            return $this->render('yaml_file/corbeilleRepertoire.html.twig', [
+                'repertoires' => $repertoires,
+                'yamlFiles'   => $yamlFiles,
+            ]);
+        }
     }
 
     #[IsGranted('REP_EDIT', subject: 'repertoire')]
@@ -91,9 +112,6 @@ final class RepertoireController extends AbstractController
         EntityManagerInterface $em
     ): JsonResponse
     {
-
-        $utilisateur = $this->getUser();
-
         $submittedToken = $request->getPayload()->get('_token');
         if (!$this->isCsrfTokenValid('delete_repertoire' . $repertoire->getId(), $submittedToken)) {
             return new JsonResponse(null, Response::HTTP_FORBIDDEN);
@@ -106,40 +124,99 @@ final class RepertoireController extends AbstractController
         return new JsonResponse(null, Response::HTTP_NO_CONTENT);
     }
 
+    #[IsGranted('GROUPE_EDIT', subject: 'groupe')]
+    #[Route('/groupe/{id}/repertoire/supprimerCorbeille/{repertoireId}', name: 'deleteRepertoireGroupe', options: ["expose" => true], methods: ['DELETE'])]
+    public function supprimerRepertoireGroupe(
+        Groupe $groupe,
+        RepertoireRepository $repertoireRepository,
+        Request $request,
+        EntityManagerInterface $em,
+        int $repertoireId
+    ): JsonResponse
+    {
+        $repertoire = $repertoireRepository->find($repertoireId);
+        if (!$repertoire) {
+            return new JsonResponse(null, Response::HTTP_NOT_FOUND);
+        }
+
+        $submittedToken = $request->getPayload()->get('_token');
+        if (!$this->isCsrfTokenValid('delete_repertoire_groupe_' . $groupe->getId() . '_' . $repertoire->getId(), $submittedToken)) {
+            return new JsonResponse(null, Response::HTTP_FORBIDDEN);
+        }
+
+        $repertoire->softDeleteForGroupe($groupe);
+
+        $em->persist($repertoire);
+        $em->flush();
+
+        return new JsonResponse(null, Response::HTTP_NO_CONTENT);
+    }
 
     #[IsGranted('REP_EDIT', subject: 'repertoire')]
     #[Route('/repertoire/restaurer/{id}', name: 'restore_repertoire')]
-    public function restore(Repertoire $repertoire, EntityManagerInterface $em): Response
+    #[Route('/groupe/{idGroupe}/repertoire/restaurer/{id}', name: 'restore_repertoire_groupe')]
+    public function restore(Repertoire $repertoire, EntityManagerInterface $em, ?int $idGroupe = null): Response
     {
         $repertoire->restore();
         $em->flush();
 
         $this->addFlash("success", "Le répertoire à bien été restauré");
-        return $this->redirectToRoute('repertoire_corbeille');
+
+        if ($idGroupe === null) {
+            return $this->redirectToRoute('repertoire_corbeille');
+        }
+
+        $groupe = $em->getRepository(Groupe::class)->find($idGroupe);
+
+        return  $this->redirectToRoute('repertoire_corbeille_groupe', ['id' => $groupe->getId()]);
     }
 
 
     #[IsGranted('REP_EDIT', subject: 'repertoire')]
     #[Route('/repertoire/supprimer/{id}', name: 'delete_repertoire_permanent')]
-    public function supprime(Repertoire $repertoire, EntityManagerInterface $em, RepertoireService $repertoireService): Response
+    #[Route('/groupe/{idGroupe}/repertoire/supprimer/{id}', name: 'delete_repertoire_permanent_groupe')]
+    public function supprime(Repertoire $repertoire, EntityManagerInterface $em, RepertoireService $repertoireService, ?int $idGroupe = null): Response
     {
         $repertoireService->deleteRepertoireWithFiles($repertoire, $em);
         $em->flush();
 
         $this->addFlash("success", "Le répertoire à bien été supprimé");
-        return $this->redirectToRoute('repertoire_corbeille');
+
+        if ($idGroupe === null) {
+            return $this->redirectToRoute('repertoire_corbeille');
+        }
+
+        $groupe = $em->getRepository(Groupe::class)->find($idGroupe);
+
+        return  $this->redirectToRoute('repertoire_corbeille_groupe', ['id' => $groupe->getId()]);
     }
 
 
     #[IsGranted('ROLE_USER')]
     #[Route('/repertoire/corbeille/delete-all', name: 'delete_repertoire_permanent_all')]
+    #[Route('/groupe/{id}/repertoire/corbeille/delete-all', name: 'delete_repertoire_permanent_all_groupe')]
     public function deleteAll(
         EntityManagerInterface $em,
         RepertoireRepository $repo,
-        RepertoireService $repertoireService
+        RepertoireService $repertoireService,
+        ?Groupe $groupe = null
     ): Response {
         $user = $this->getUser();
-        $repertoires = $repo->findDeletedByUser($user);
+
+        if ($groupe === null) {
+            $repertoires = $repo->findDeletedByUser($user);
+            foreach ($repertoires as $rep) {
+                $repertoireService->deleteRepertoireWithFiles($rep, $em);
+            }
+            $em->flush();
+
+            $this->addFlash('success', 'Tous vos répertoires supprimés définitivement !');
+            return $this->redirectToRoute('repertoire_corbeille');
+        }
+
+        $this->denyAccessUnlessGranted('GROUPE_EDIT', $groupe);
+
+        $repertoires = $repo->findDeletedByGroupe($groupe);
 
         foreach ($repertoires as $rep) {
             $repertoireService->deleteRepertoireWithFiles($rep, $em);
@@ -147,10 +224,11 @@ final class RepertoireController extends AbstractController
 
         $em->flush();
 
-        $this->addFlash('success', 'Tous les répertoires supprimés définitivement !');
+        $this->addFlash('success', 'Tous les répertoires du groupe supprimés définitivement !');
 
-        return $this->redirectToRoute('repertoire_corbeille');
+        return $this->redirectToRoute('repertoire_corbeille_groupe', ['id' => $groupe->getId()]);
     }
+
 
     /**
      * @throws Exception
